assessment/CrewAI Task/requirements.txt
crewai==0.28.9rc2
langchain==0.1.16
langchain-google-genai==0.0.6
google-generativeai==0.3.2
python-dotenv>=1.0.1
pydantic>=2.4.2
langsmith==0.1.18
streamlit
